fixing uno test

+++ nanopay/src/net/nanopay/security/test/MerkleTreeTest.js
@@ -33,7 +33,7 @@ foam.CLASS({
             }
 
             @Override
-            protected int computeTotalTreeNodes() {
+            protected double computeTotalTreeNodes() {
               return super.computeTotalTreeNodes();
             }
           }
